n
Add stub for PatternBuilder in MainForm switch

A new case for "PatternBuilder" has been added to the switch statement in the MainForm class. Currently, this case does not instantiate a new game or perform any actions; it simply exits the switch. This addition suggests a placeholder for future development or external handling of the "PatternBuilder" game type.